refactor: Replace CategoryDtoRef with CategoryDto and update related mappings in CategoryController, CategoryService, and CategoryMapper
