+++ src/main/java/com/regisx001/blog/controllers/CategoryController.java
@@ -4,12 +4,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.regisx001.blog.domain.dto.CategoryDto;
-import com.regisx001.blog.domain.dto.CategoryDtoRef;
-import com.regisx001.blog.domain.dto.requests.CreateCategoryRequest;
 import com.regisx001.blog.domain.dto.requests.UpdateCategoryRequest;
 import com.regisx001.blog.domain.entities.Category;
 import com.regisx001.blog.mappers.CategoryMapper;
-import com.regisx001.blog.mappers.CategoryMapperRef;
 import com.regisx001.blog.services.CategoryService;
 
 import jakarta.validation.Valid;
@@ -37,18 +34,17 @@ public class CategoryController {
 
     private final CategoryService categoryService;
     private final CategoryMapper categoryMapper;
-    private final CategoryMapperRef categoryMapperRef;
 
     @GetMapping
-    public ResponseEntity<Page<CategoryDtoRef.Detailed>> getAllCategories(Pageable pageable) {
-        Page<CategoryDtoRef.Detailed> categories = categoryService.getAllCategories(pageable);
+    public ResponseEntity<Page<CategoryDto.Detailed>> getAllCategories(Pageable pageable) {
+        Page<CategoryDto.Detailed> categories = categoryService.getAllCategories(pageable);
         return ResponseEntity.ok(categories);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<CategoryDtoRef.Detailed> getCategoryById(@PathVariable UUID id) {
+    public ResponseEntity<CategoryDto.Detailed> getCategoryById(@PathVariable UUID id) {
         Category category = categoryService.getCategoryById(id);
-        return ResponseEntity.ok(categoryMapperRef.toDetailedDto(category));
+        return ResponseEntity.ok(categoryMapper.toDetailedDto(category));
     }
 
     // @PostMapping
@@ -60,10 +56,10 @@ public class CategoryController {
     // }
 
     @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<CategoryDtoRef.Detailed> createCategory(
-            @Valid @ModelAttribute CategoryDtoRef.CreateWithImageRequest categoryRequest) {
+    public ResponseEntity<CategoryDto.Detailed> createCategory(
+            @Valid @ModelAttribute CategoryDto.CreateWithImageRequest categoryRequest) {
         Category category = categoryService.createCategory(categoryRequest);
-        return new ResponseEntity<>(categoryMapperRef.toDetailedDto(category), HttpStatus.CREATED);
+        return new ResponseEntity<>(categoryMapper.toDetailedDto(category), HttpStatus.CREATED);
     }
 
     // TODO: IMPLEMENT LATER
