+++ src/main/java/com/regisx001/blog/services/impl/CategoryServiceImpl.java
@@ -7,12 +7,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import com.regisx001.blog.domain.dto.CategoryDto;
-import com.regisx001.blog.domain.dto.CategoryDtoRef;
-import com.regisx001.blog.domain.dto.requests.CreateCategoryRequest;
 import com.regisx001.blog.domain.dto.requests.UpdateCategoryRequest;
 import com.regisx001.blog.domain.entities.Category;
 import com.regisx001.blog.mappers.CategoryMapper;
-import com.regisx001.blog.mappers.CategoryMapperRef;
 import com.regisx001.blog.repositories.CategoryRepository;
 import com.regisx001.blog.services.CategoryService;
 import com.regisx001.blog.services.StorageService;
@@ -26,15 +23,14 @@ public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
     private final StorageService storageService;
     private final CategoryMapper categoryMapper;
-    private final CategoryMapperRef categoryMapperRef;
 
     @Override
-    public Page<CategoryDtoRef.Detailed> getAllCategories(Pageable pageable) {
-        return categoryRepository.findAll(pageable).map(categoryMapperRef::toDetailedDto);
+    public Page<CategoryDto.Detailed> getAllCategories(Pageable pageable) {
+        return categoryRepository.findAll(pageable).map(categoryMapper::toDetailedDto);
     }
 
     @Override
-    public Category createCategory(CategoryDtoRef.CreateWithImageRequest categoryRequest) {
+    public Category createCategory(CategoryDto.CreateWithImageRequest categoryRequest) {
         String imagePath = null;
 
         if (categoryRequest.image() != null) {
